DEXA-41 Code refactor
